Test request: Create/Set/Get IO ports
- Fix ITEM_NOT_FOUND data type (using int)
- Create remove and removeAll IOs
- Handling Create/Set/Get Clear IO ports
- Create Test request test cases
- Reading debug responses

+++ lib/Manager/Manager.h
@@ -39,8 +39,8 @@ class Manager
         unsigned int totalWaterTanks = 0;
         unsigned int totalWaterSources = 0;
 
-        long getWaterTankIndex(String name);
-        long getWaterSourceIndex(String name);
+        int getWaterTankIndex(String name);
+        int getWaterSourceIndex(String name);
 };
 
 #endif
